Port yuzu-emu/yuzu#5158: "video_core: Remove unnecessary enum class casting in logging messages"

+++ src/video_core/regs_framebuffer.h
@@ -200,7 +200,7 @@ struct FramebufferRegs {
         case ColorFormat::RGBA4:
             return 2;
         default:
-            LOG_CRITICAL(HW_GPU, "Unknown color format {}", static_cast<u32>(format));
+            LOG_CRITICAL(HW_GPU, "Unknown color format {}", format);
             UNIMPLEMENTED();
         }
     }
@@ -265,7 +265,7 @@ struct FramebufferRegs {
             return 4;
         }
 
-        ASSERT_MSG(false, "Unknown depth format {}", static_cast<u32>(format));
+        ASSERT_MSG(false, "Unknown depth format {}", format);
     }
 
     // Returns the number of bits per depth component of the specified depth format
@@ -278,7 +278,7 @@ struct FramebufferRegs {
             return 24;
         }
 
-        ASSERT_MSG(false, "Unknown depth format {}", static_cast<u32>(format));
+        ASSERT_MSG(false, "Unknown depth format {}", format);
     }
 
     INSERT_PADDING_WORDS(0x10); // Gas related registers
